refactor: Update naming and useEffect usage
Rename 'const' to 'consts' to enhance naming consistency throughout the
codebase, and eliminate redundant 'useEffect' invocation in Loadable component.
