+++ components/icon/svgIcon/index.tsx
@@ -1,4 +1,4 @@
-import { VIEWBOX } from '@icon/icon.const';
+import { VIEWBOX } from '@icon/icon.consts';
 import { TypesPropsOptionsSvg } from '@icon/icon.types';
 import { memo } from 'react';
