+++ components/label/label.consts.ts
